chore: implement hooks module in core

Former-commit-id: 4a18412b8dbb8a6d39e2451cb0c82f810eddef0e

+++ pkg/core/app/app.go
@@ -22,7 +22,7 @@ import (
 	"time"
 )
 
-func NewApp(logger *zap.Logger, id int, cmd string) core.App {
+func NewApp(logger *zap.Logger, id uint64, cmd string) core.App {
 	app := &App{
 		logger: logger,
 		id:     id,
@@ -35,7 +35,7 @@ func NewApp(logger *zap.Logger, id int, cmd string) core.App {
 type App struct {
 	logger           *zap.Logger
 	docker           docker.Client
-	id               int
+	id               uint64
 	cmd              string
 	kind             utils.CmdType
 	containerDelay   time.Duration
